requirements.txt
asgiref==3.5.2
astroid==2.11.5
attrs==21.4.0
Automat==20.2.0
autopep8==1.6.0
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
colorama==0.4.4
constantly==15.1.0
cryptography==37.0.2
cssselect==1.1.0
cycler==0.11.0
dill==0.3.5.1
Django==4.0.5
django-crispy-forms==1.14.0
django-filter==22.1
filelock==3.7.1
fonttools==4.33.3
gunicorn==20.1.0
httplib2==0.20.4
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
isort==5.10.1
itemadapter==0.6.0
itemloaders==1.0.4
jmespath==1.0.1
kiwisolver==1.4.3
lazy-object-proxy==1.7.1
lxml==4.9.0
matplotlib==3.5.2
mccabe==0.7.0
numpy==1.22.4
packaging==21.3
parsel==1.6.0
Pillow==9.1.1
platformdirs==2.5.2
Protego==0.2.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyDispatcher==2.0.5
pylint==2.14.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
Scrapy==2.6.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tldextract==3.3.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
Twisted==22.4.0
twisted-iocpsupport==1.0.2
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
w3lib==1.22.0
whitenoise==6.2.0
wrapt==1.14.1
zope.interface==5.4.0
